refactor: extract overview HeaderSection component
